Refactor server code to leverage new imports
