+++ test/integration/browser/image.ts
@@ -8,7 +8,7 @@ import Upscaler from 'upscaler';
 import fs from 'fs';
 import path from 'path';
 import type { Page } from 'puppeteer';
-import { BrowserTestRunner } from '@internals/test-runner/browser';
+import { BrowserTestRunner } from '../../../internals/test-runner/src/browser';
 import { MODELS_DIR } from '../../../scripts/package-scripts/utils/constants';
 
 const PIXEL_UPSAMPLER_DIR = path.resolve(MODELS_DIR, 'pixel-upsampler/test/__fixtures__');
