+++ test/integration/browser/model.ts
@@ -7,7 +7,7 @@ import Upscaler, { ModelDefinition } from 'upscaler';
 import * as tf from '@tensorflow/tfjs';
 import type { Tensor3D, } from '@tensorflow/tfjs';
 import * as tfn from '@tensorflow/tfjs-node';
-import { BrowserTestRunner } from '@internals/test-runner/browser';
+import { BrowserTestRunner } from '../../../internals/test-runner/src/browser';
 import path from 'path';
 import { MODELS_DIR } from '../../../scripts/package-scripts/utils/constants';
