+++ test/integration/model/model.ts
@@ -7,8 +7,8 @@ import { DIST as UMD_DIST, mockCDN as umdMockCDN } from '../../lib/umd/prepare';
 import Upscaler, { ModelDefinition } from 'upscaler';
 import * as tf from '@tensorflow/tfjs';
 import { AvailableModel, getFilteredModels } from '../../../scripts/package-scripts/utils/getAllAvailableModels';
-import { BrowserTestRunner } from '@internals/test-runner/browser';
-import { NodeTestRunner, Main } from '@internals/test-runner/node';
+import { BrowserTestRunner } from '../../../internals/test-runner/src/browser';
+import { NodeTestRunner, Main } from '../../../internals/test-runner/src/node';
 import path from 'path';
 import { MODELS_DIR, TMP_DIR } from '../../../scripts/package-scripts/utils/constants';
 import { getPackageJSON } from '../../../scripts/package-scripts/utils/packages';
